refactor(agent-test): share hardened local adapter server
